+++ lib/domain/model/search/server_side_search_categories.dart
@@ -20,7 +20,6 @@ class ServerSideSearchCategories with EquatableMixin {
         groupings: Groupings(
           groupBy: [Group(key: GroupKey.roomId)],
         ),
-        keys: [KeyKind.contentBody],
         orderBy: SearchOrder.recent,
         filter: searchFilter,
       ),