+++ lib/pages/chat/events/message/message_content_with_timestamp_builder.dart
@@ -138,8 +138,10 @@ class MessageContentWithTimestampBuilder extends StatelessWidget {
                               Column(
                             crossAxisAlignment: CrossAxisAlignment.start,
                             children: [
-                              event.hideDisplayName(nextEvent) ||
-                                      event.hideDisplayNameInBubbleChat
+                              event.hideDisplayName(
+                                nextEvent,
+                                responsiveUtils.isMobile(context),
+                              )
                                   ? const SizedBox()
                                   : DisplayNameWidget(
                                       event: event,